+++ CHANGELOG.md
@@ -29,6 +29,7 @@
     instead.
 * MISC
   * Removed hardcoded Graph urls and replaced by MSCloudLoginAssistant values.
+  * Add separate module handling for PowerShell Core.
 * DEPENDENCIES
   * Updated DSCParser to version 2.0.0.14.
   * Updated Microsoft.Graph to version 2.25.0.